# api/openapi.yaml
openapi: 3.1.0
info:
  title: LyraNest Community API
  version: 0.3.0
  description: Authenticated API for library discovery, playback, management, normalized sidecar lyrics, and persistent per-user listening data.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/tracks:
    get:
      operationId: listTracks
      summary: List scanned tracks
      responses:
        "200":
          description: Scanned tracks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackListResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/library/scan:
    post:
      operationId: scanLibrary
      summary: Rescan the configured music library
      description: Synchronously scans the library directory and replaces the in-memory track list when successful.
      responses:
        "200":
          description: Library scan completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryScanResponse"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current server state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    TrackListResponse:
      type: object
      required: [tracks, total]
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
          minimum: 0
    LibraryScanResponse:
      type: object
      required: [tracks, total, scanned_at]
      properties:
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        total:
          type: integer
          minimum: 0
        scanned_at:
          type: string
          format: date-time
...
    Track:
      type: object
      required: [id, title, file_name, extension, size_bytes, modified]
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        limit:
          type: integer
          minimum: 1
          maximum: 25
          default: 15
          description: Maximum number of provider candidates to search for.
        file_name:
          type: string
        extension:
          type: string
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        modified:
          type: string
          format: date-time
        artwork_url:
          type: string
          format: uri
          description: Optional explicit artwork URL. Clients may use the track artwork endpoint when omitted.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string